+++ lib/webpacker/react/railtie.rb
@@ -12,7 +12,8 @@ module Webpacker
       initializer :webpacker_react_renderer, group: :all do |_app|
         ActionController::Renderers.add :react_component do |component_name, options|
           props = options.fetch(:props, {})
-          html = Webpacker::React::Component.new(component_name).render(props)
+          tag_options = options.fetch(:tag_options, {})
+          html = Webpacker::React::Component.new(component_name).render(props, tag_options)
           render_options = options.merge(inline: html)
           render(render_options)
         end