+++ test/example_app/app/controllers/custom_tag_controller.rb
@@ -1,0 +1,8 @@
+class CustomTagController < ApplicationController
+  def view_component
+  end
+
+  def controller_component
+    render react_component: "HelloReact", props: { name: "a component rendered from a controller in a span" }, tag_options: { tag: :span }
+  end
+end